# fcr-backend-app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "FCR Rest API Documentation",
    "description": "Backend Server Documentation Rest APIs",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "http://185.192.96.18:31714",
  "basePath": "/",
...
  "schemes": ["http","https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/gif": {
      "get": {
        "tags": ["Gif"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/logo": {
      "get": {
        "tags": ["Logo"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/meetings/": {
      "post": {
        "tags": ["Meetings"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "example": "any"
                },
                "eventType": {
                  "example": "any"
                },
                "coachName": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "habillements": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Meetings"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/subscribe/addSubscriber": {
      "post": {
        "tags": ["Subscribe"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}